+++ lib/svea_payments/base.rb
@@ -4,14 +4,27 @@ require 'nokogiri'
 
 module SveaPayments
   module Base
-    def send_post_request(uri, form_data, token)
+    def send_post_request(uri, form_data, token, strict: false)
       request = Net::HTTP::Post.new(uri)
       request.content_type = 'application/x-www-form-urlencoded'
       request.body = form_data
       request['Authorization'] = token
 
-      response = Net::HTTP.start(uri.hostname, uri.port, use_ssl: uri.scheme == 'https') do |http|
+      options = { use_ssl: uri.scheme == 'https' }
+      options.merge!(open_timeout: 10, read_timeout: 30, write_timeout: 30) if strict
+      response = Net::HTTP.start(uri.hostname, uri.port, **options) do |http|
+        http.max_retries = 0 if strict
         http.request(request)
+      end
+
+      if strict
+        raise SveaPayments::HTTPError, response.code unless response.is_a?(Net::HTTPSuccess)
+
+        begin
+          return Nokogiri::XML(response.body) { |config| config.strict.nonet }
+        rescue Nokogiri::XML::SyntaxError
+          raise SveaPayments::InvalidResponseError, 'Invalid XML response; request outcome may be unknown'
+        end
       end
 
       Nokogiri::XML(response.body)